third_party/WebKit/LayoutTests/inspector/debugger/script-formatter-breakpoints-expected.txt

Tests the script formatting is working fine with breakpoints.

Debugger was enabled.

Running: testBreakpointsInOriginalAndFormattedSource
Script execution paused.
Breakpoint sidebar pane while paused in raw

    script-formatter-breakpoints.html:12var f=0;
Script execution resumed.
Script execution paused.
Breakpoint sidebar pane while paused in pretty printed

    script-formatter-breakpoints.html:17var f = 0;
Breakpoint sidebar pane while paused in raw

    script-formatter-breakpoints.html:12var f=0;
Script execution resumed.

Running: testBreakpointSetInOriginalAndRemovedInFormatted
Adding breakpoint.
Formatting.
Removing breakpoint.
Unformatting.

Running: testBreakpointsSetAndRemoveInFormattedSource
Script execution paused.
Breakpoint sidebar pane while paused in pretty printed

    unformatted.js:4var b = 1;
Breakpoint sidebar pane while paused in raw

Script execution resumed.

Running: testBreakpointsSetInFormattedAndRemoveInOriginalSource
Script execution paused.
Breakpoint sidebar pane while paused in pretty printed

    unformatted.js:4var b = 1;
Breakpoint sidebar pane while paused in raw

    unformatted.js:3var b=1;var c=3;var d=4;var e=5;
Breakpoint sidebar pane after removed

Script execution resumed.
Debugger was disabled.
